Backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.1.0
openpyxl==3.1.2
werkzeug==2.3.7
requests==2.31.0
python-dotenv==1.0.0
numpy==1.24.3

# Enhanced Education Consultant System Dependencies
sentence-transformers==2.2.2
faiss-cpu==1.7.4
scikit-learn==1.3.0
transformers==4.33.2
torch==2.0.1
torchvision==0.15.2
tokenizers==0.13.3

# Optional: For better performance with GPU
# faiss-gpu==1.7.4  # Uncomment if you have CUDA GPU support 